Remove special caching of JAVA_LANG_OBJECT in TextTypeParser
Previously, the `obtainObjectType()` would access the cache directly
and if an entry could not be found would create a `TextClassTypeItem`
directly. The problem with that is that there were two paths to get
a `ClassTypeItem` for `java.lang.Object`, through `cachedParseType()`
and through this method and each path would produce a different
result. The first would return a type with either `PLATFORM` or
`NONNULL` depending on the `kotlinStyleNulls` setting. The second
would always return a type with `PLATFORM`.

That would lead to weird behavior that was dependent on the order in
which those paths were taken. This change removes the second path so
they both use `cachedParseType()`.

Bug: 323614314
Test: ./gradlew
Change-Id: I652fa5f55df1b08507f95932c4d383a7e19985e4