Add ExceptionTypeItem and ThrowableType wrapper around it
Previously, all `ThrowableType` instances had to be resolved to a
`ClassItem` (or a `TypeParameterItem` but that is not an issue here).
That makes constructing a codebase difficult as it has to support
reentrant `ClassItem` construction.

This change adds a new `ExceptionTypeItem` interface which can be
either a `ClassTypeItem` or a `VariableTypeItem`, and a `ThrowableType`
factory method to create a `ThrowableType` instance.

It also creates a new `ThrowableType` wrapper class around a
`ClassTypeItem` (it does not need one for `VariableTypeItem` as it
can reuse the existing wrapper around `TypeParameterItem`).

One of the tests was broken because it relied on exception classes
being resolved immediately. That had to be restructured. The order
of exceptions in the throws list had to be changed as that relied on
the order in which classes were resolved.

The test is now broken on Psi because it has similar issues to the ones
being fixed for Turbine (unsurprising given Turbine does a similar job
to Psi). They will be fixed in a follow up change.

Bug: 323678811
Test: ./gradlew
Change-Id: I91e29d98696cffb5bcf3cc03917ac696287b6bb6