Reuse TypeParameterScopes for ClassItems
Previously, a new `TypeParameterScope` (albeit using the `ClassItem`'s
type parameters) would be created for outer classes and previously
created classes, created in a previously processed file. Apart from the
slight duplication that could cause a serious problem where a type is
parsed using a newly created list of type parameters but actually it
should be referencing the existing class's type parameters.

That already happens with interface types. When interface types are
parsed they use the newly created type parameters and not the existing
class's type parameters. That is not actually an issue at the moment
because the interface types are discarded if there is an existing
class. However, similar issues would arise when the super class types
are parsed in the same way.

e.g. Given the following:

file1.txt:

    // Signature format: 2.0
    package test.pkg {
      public class PublicBase<T> {
      }
      public class Generic<T> extends PublicBase<T> {
        field public T publicField;
      }
    }

file2.txt:

    // Signature format: 2.0
    package test.pkg {
      public class SystemBase<T> extends PublicBase<T> {
      }
      public class Generic<T> extends SystemBase<T> {
        field public T systemField;
      }
    }

Assuming super class types were parsed alongside interface types then
without this change the `T` in `SystemBase<T>` would reference a
different type parameter to `T` in `Generic<T>` from `file1.txt`.

This change avoids that by reusing the `TypeParameterScope` that was
previously created during parsing of the first class definition.

Bug: 323614314
Test: ./gradlew
Change-Id: I392992cb5557402ecf25c057a89af2dfb1d8fa80