Make TextCodebaseBuilder copy superClassType()
Previously, the `TextCodebaseBuilder` would not copy the
`superClassType()` from a `TextClassItem` when building a new
`TextCodebase`. Instead it would just use the `ReferenceResolver` to
add the default super class based on the `ClassKind`. This change
removes the use of `ReferenceResolver` and simply copies the
`superClassType()` over instead.

This is a fix to allow progress on the `superClassType()` work. The
other issues with `TextCodebaseBuilder`, e.g. not copying interfaces
either will be rectified in future work.

Bug: 322458348
Test: ./gradlew
Change-Id: I6c2e781a24ac90902969810ae988d9b6109bb742