Cache types in the best TypeParameterScope
Previously, types would be cached in the `TypeParameterScope` within
which they were being parsed. However, that is not the best (in the
sense of maximizing caching) one to use. The best one to use is the
closest enclosing scope that provides at least one type parameter
whose name is used in the type.

```
public class Outer<T,U> {
  public class Middle<T> {
    public class Inner<V> {
        public V field1; // This should use Inner<V>
	public T field2; // This should use Middle<T>
	public U field3; // This should use Outer<T,U>
	public int field4; // This should use the global scope (empty)
    }
  }
}
```

This change modifies the caching to use the best scope.

Bug: 323614314
Test: ./gradlew
Change-Id: I7f48fdea153023824f50540ef20fd0e76fe9ced2