Modify test for PsiImmediateClassType workaround to reveal issue
The test was added in https://r.android.com/2883447 to verify a
workaround for an issue with `PsiImmediateClassType`. Unfortunately,
the test was only working because of a happy coincidence.

The workaround is supposed to handle the case where a type argument is
missing from the `PsiType`, it tries to fix that by resolving the
`PsiType` to a `PsiClass`, getting its type parameters and then
mapping it to the type arguments and using that.

The problem is that the resulting type arguments are incorrect. e.g.
the test looks something like this:
```
public class Foo<E> ... {
     ...Collection<E>...
}
```

If the `PsiType` with the issue is `Collection<E>` (so the `E` is
missing) then it will resolve the class to `Collection`, gets its
type parameter which is `E` and use that, so the resulting type will
be `Collection<E>`. It looks right but is actually wrong because the
`E` is referring to the type parameter from `Collection`, not `Foo`.

The test was modified (and `PsiTypeParameterItem.toString()`
implemented) to make that clearer by renaming the `E` to `Z`. Now the
test is failing because the `E` from `Collection<E>` no longer matches
`Z` from `Foo<Z>`.

This change disables the test so that the follow up work (which first
detected this issue) can be submitted.

Bug: 324064238
Test: ./gradlew
Change-Id: Ib311da8dd4c09dabe002f1b1a61e94d2f576c457