Stop creating PsiTypeParameterItems lazily
Previously, the `PsiTypeParameterList`s for `PsiClassItem`,
`PsiMethodItem` and `PsiConstructorItem` were created lazily on demand.
Also, the `PsiTypeParameterList` lazily created the
`PsiTypeParameterItem`s.

That requires complex logic (added in https://r.android.com/2952677) to
find a `PsiTypeParameterItem` instead of just creating a map from
`PsiTypeParameter` to `PsiTypeParameterItem` and populating it when
`PsiTypeParameterItem`s are created.

This change creates the `PsiTypeParameterList` and its
`PsiTypeParameterItem`s early and passes it into the constructors of
the above `Psi*Item` classes.

Bug: 324064238
Test: ./gradlew
Change-Id: I19d9cd51ab9157baa1c538d563ade5b2aba72563