Extract common code for creating a List<TypeParameterItem>
Refactors the existing Turbine code into two separate parts. A generic
common method for use by all models that is parameterized by the model
specific types that represent parameters and their bounds.

This might not seem like much of an improvement on its own but it is
a step towards deduping model specific implementations of
`TypeParameterItem` and `TypeParameterList`.

Bug: 324196754
Test: ./gradlew
Change-Id: Id241488407f3e138682d03afc6d9fd97e06589d2