Remove DefaultTypeItem.codebase
There was only one use for it in `DefaultTypeItem` and that was to pass
it to `TypeStringConfiguration` which only used it to resolve an
`AnnotationItem` to a class. However `AnnotationItem` had its own
`codebase` property so that was used instead.

The sub-classes of `ClassTypeItem` all need access to it to resolve a
`ClassItem` so they just have their own property.

Also, the `PsiVariableTypeItem` also needs access to it, at least for
the moment so it has its own property too.

Bug: 324196754
Test: ./gradlew
Change-Id: I98f31c19ef36926a0f20b2b7f25ac40204191bf0