Simplify the creation of inner classes
Previously, the inner classes used `findOrCreateClass(ClassSymbol)` to
create the `TurbineClassItem` for inner classes. That was needed
because when an inner class extended another that was defined later in
the class then the later one would need to be defined first. That was
because the super class  was resolved immediately. Unfortunately, it
created very complex reentrant code that was difficult to reason about
and would not easily support using `TypeParameterScope` to track type
parameter variables from outer classes on inner classes.

However, the preceding change that defers resolving super class means
that is no longer needed so this simplifies the inner class creation
by calling `createClass(ClassSymbol)` directly and simplifies the
`findOrCreateClass(ClassSymbol)` method.

Bug: 324317136
Test: ./gradlew
Change-Id: Iab89b81b4c04859284ea0447ba0a30dd18fc4109