Defer resolution of ThrowableType in text model
Previously, the text model stored the names of the throws types in the
`TextMethodItem` during parsing for use during reference resolving. At
which point it resolved them to a `ClassItem` or `TypeParameterItem` and
then wrapped that in the appropriate `ThrowableType`. This change
avoids that by creating an `ExceptionTypeItem`, wrapping it in a
`ThrowableType` and storing it in the `TextMethodItem` directly during
parsing.

The previous resolution code also made sure that if the throws type
did not resolve to an actual class and so required a stub that the
stub extended `java.lang.Throwable` (unless it was
`java.lang.Throwable`). This change does the same by having
`TextTypeItemFactory.getExceptionType(...)` record that the type
requires a `StubKind.THROWABLE` so if a stub is created for it then
it will have the correct super class.

As resolving the throws list was the last of the work done by the
`ReferenceResolver` this change removes that whole class.

Bug: 323678811
Test: ./gradlew
Change-Id: I99b157e8b342686e55780470c427818f96f8107d