Construct a generic PsiType for a PsiClass instead of a raw one
Previously, when run on a class `Foo<X>` `getType(PsiClass)` would
create a raw type `Foo` (with no type arguments). It then relied on the
workaround in `PsiTypeItem.computeTypeArguments(...)` to add the
generic type arguments back in to get `Foo<X>`.

It appears that workaround was originally added in for this issue (or
something similar) but was then triggered by a bug in Psi. It sort of
worked in that case, by fluke. Now, the Psi bug needs working around
properly and fixing that broken a default constructor test that relied
on the previous behavior to add in the missing type arguments.

This change fixes the original problem properly and creates a proper
generic `PsiType` in the first place so it no longer relies on the
workaround. Which means fixing that workaround will not break the test.

Bug: 324317136
Test: ./gradlew
Change-Id: If4ba41a612eae0aa226a21256fdcae078886444a