Defer resolution of super classes and interfaces in psi model
Following on from similar changes in the Turbine and Text model. This
change greatly simplifies the initialization of a `PsiClassItem`'s
super class and interfaces by using `ClassTypeItem` and deferring the
resolution until needed.

As with similar change this broke a test that relied on classes that
are not in the codebase being tested but which are super classes or
interfaces of ones that are being available in the codebase. This
change adds a new `assertResolvedClass(...)` assertion that will
force the class to be resolved.

Bug: 324317136
Test: ./gradlew
Change-Id: I1a677de0293597c5a286c63366641cb355ab794c