Allow ThrowableType.classItem to be null
Currently, the Psi model resolves throws classes and if it cannot find
a class then it is skipped and then a `ThrowableType` is created around
it. However, there is currently no way to have a `ThrowableType`
indicate that it could not be resolved for some reason. After this
change that is done by having `classItem` be `null`.

Adds a test, which does not yet work on Psi but that will be fixed in
some follow up changes.

Bug: 323678811
Test: ./gradlew
Change-Id: I5a6caf6ed305a15e14caa8817794c7fd83576686