Prevent registering duplicate PsiClassItems
Previously, the code did not check that multiple `PsiClassItem`s were
being registered under the same qualified name. It relied on the code
following strict rules to avoid that, with appropriate checks to find
the class before creating one. However, while that worked it left it
fragile to refactorings that might accidentally break the rules.

This change adds a simple layer of protection to detect the problem
early.

Bug: 324317136
Test: ./gradlew
Change-Id: I748a2b3dcb0366dfe85715218551232358125110