Stop calling finishInitialization on inner classes twice
Previously, when an inner class was created when `initializing` was
`false` it could have its `PsiClassItem.finishInitialization()` twice.
Once from `createClass(...)` when it was created and again from the
`finishInitialization()` method of its containing class.

This change avoids that by only calling `finishInitialization` on
the outer most class created by `findOrCreateClass(...)`.

Bug: 324317136
Test: ./gradlew
Change-Id: I19bbd830951bedd720d9c657c11d4de9c4e4538c