Fix broken MultipleCompatibilityFilesTest tests
The test does not use fully qualified names in the `throws` clause
which means it does not use the classes that it expects and instead
uses stubs for `java.lang.Baz` and `java.lang.Bar`. That reports the
wrong issues because while it was set up for `test.pkg.Baz` to extend
`test.pkg.Bar` that is not true of `java.lang.Baz` and `java.lang.Bar`.

This change fully qualifies the reference and updates the now, correct
issues.

Bug: 323678811
Test: ./gradlew
Change-Id: I5ada313655c20c5b77a3f5d800aa4066b9fc9c77