Support TypeParameterScope in PsiTypeItemFactory
Adds support for `TypeParameterScope` in `PsiTypeItemFactory` and
updates the use of it within the psi model to match the different
scopes and correctly track all the different `PsiTypeParameterItem`s.
However, the scope is not actually used to resolve references to those
items within `PsiType`s. That will be done in a follow-up to keep the
changes easier to review.

Bug: 323678811
Test: ./gradlew
Change-Id: Iaa936068a49a1e2e66b673a2c080506db86c8b61