Prepare for replacing ThrowableType with ExceptionTypeItem
The `ThrowableType` was only ever an intermediate step between what
there was before and `ExceptionTypeItem`. This change prepares for
replacing it with the latter by adding some specialist implementation
classes that implement `ClassTypeItem` and `VariableTypeItem` by
delegating to the supplied property. That allows the `isTypeParameter`
property to be replaced with a check for `VariableTypeItem` as will
need to be done when using `ExceptionTypeItem`.

This purposely does not have `ThrowableType` extend `ExceptionTypeItem`
itself because that would make the existing `when(exceptionTypeItem)`
statements non-exhaustive.

Bug: 323678811
Test: ./gradlew
Change-Id: I376f2270ba5331e1e907de09362c4a3451a367c4