Make TypeParameterList a List<TypeParameterItem>
Most uses of `TypeParameterList` need a `List<TypeParameterItem>` so
rather than require them all to call `typeParameters()` to get the list
this makes `TypeParameterList` extend `List<TypeParameterItem>` and
implement it via delegation. It also replaces `typeParameterCount()`
with `size`.

The `equals(Object)` and `hashCode()` methods were implemented explicitly.

The `TypeParameterList` cannot be replaced entirely by a
`List<TypeParameterItem>` because it has its own `toString()` method
but this will make it simpler to use.

Bug: 324196754
Test: ./gradlew
Change-Id: I6d003bdb3704e424c8364185529a9c2814108cdf