Simplify initialization of TurbineConstructorItem.returnType
Previously, it was set to `void` (based off the underlying Turbine
type) and then fixed up later. This change initializes it to
`ClassItem.type()` just as the other models do. That required that
the `TurbineClassItem.containingClass` be set immediately after
creation to ensure that the outer class type was initialized
correctly.

Bug: 325454211
Test: ./gradlew
Change-Id: If7944fa90a598d616d9c1ea653004d389e4a3b77