Fix ClassItem.type() test for text model
This adds a new `DefaultResolvedClassTypeItem` class that implements
`ClassTypeItem` but does not need to resolve the `ClassItem` as it is
given one at construction time. A generic method is added to create
one for a `ClassItem`, setting up the `arguments` correctly where
needed. A new shared `emptyNonNullModifiers` was added that replaces
the `TextCodebase.emptyTypeModifiers`.

This also removes a call to `duplicate(TypeNullability.NONNULL)` that
is no longer needed.

Bug: 325454211
Test: ./gradlew
Change-Id: Ibea9631997880e4bdc66ce4836c144d0c5c0f7d4