Extract useful methods from ApiFile.parseType
Previously, `parseType` would scan the token stream for a type string,
parse that into a `TypeItem` and then synchronize the `TypeNullability`
of that with any nullability annotations specified on the item. This
change extracts two new methods, `scanForTypeString` and
`synchronizeNullability`, the parsing does not need its own method as
it is done by `TypeItemFactory.getGeneralType(String)`.

This will allow the scanning, parsing and synchronizing to be done
separately where necessary.

Bug: 325454211
Test: ./gradlew
Change-Id: I3206933a106ea0f4951ffecfb0864d6ead0f5cee