Add FileLocationTracker interface implemented by Tokenizer
A `Tokenizer` object consumers characters from the buffer and produces
a stream of tokens. It is inherently stateful and care has to be taken
that tokens are not consumed in the wrong order. It also tracks the
location for error messages and exceptions. This change adds a
`FileLocationTracker` interface that allows those methods that only
want to know the location to avoid being passed the whole `Tokenizer`
making them safer and easier to reason about. i.e. methods like
`parseValue`, `parsePrimitiveValue` and `ApiParseException`
constructor.

Bug: 325454211
Test: ./gradlew
Change-Id: Id651a817ee0c1e79dbd6fbb44887042ea891f093