Set nullness info on PsiClassItem.type()
Previously, the `PsiClassItem.type()` did not provide any
`KotlinTypeInfo` which meant that its `nullability` was `PLATFORM`.
However, because it was shared with the `PsiConstructorItem`'s return
type its `nullability` was set to `NONNULL` when the
`PsiConstructorItem`'s `nullability` was fixed up. As a result removing
`ConstructorItem.implicitNullness()` broke `nullability` tests for
`PsiClassItem.type()` and `PsiConstructorItem.returnType`.

This change sets the `KotlinTypeInfo` for the `PsiClassItem.type()` and
removes `ConstructorItem.implicitNullness()` as it is no longer needed.

It "broke" the `Basic Kotlin stubs` test but that is basically broken
anyway as the output is not valid Kotlin. However, the change seems
like a minor improvement.

Bug: 324196754
Test: ./gradlew
Change-Id: I2b741a70ed9e959cc06fb65195528dd676a066fe